// projects/shared-library/src/lib/core/enums/lib-global-enum.ts
export enum LocalRoleOnProject {
  CREATOR = 'CREATOR',
  AFFECTED_PROJECT_OWNER = 'AFFECTED_PROJECT_OWNER',
  EDITOR = 'EDITOR',
  VIEWER = 'VIEWER',
}

export enum IncomingRoleOnProject {
  PROJECT_OWNER = 'PROJECT_OWNER',
  EDITOR = 'EDITOR',
  VIEWER = 'VIEWER',
}

export enum UserRoles {
  ADMIN = 'ADMIN',
  PROJECT_OWNER = 'PROJECT_OWNER',
  USER = 'USER',
}

export enum Step {
  CREATE_PROJECT = 'Create Project',
  PROJECT_DETAILS = 'PROJECT_DETAILS',
  ADD_DATA = 'ADD_DATA',
  CONFIGURE_DATA = 'CONFIGURE_DATA',
  EXPLORE_DATA = 'EXPLORE_DATA',
  PROCESSORS = 'PROCESSORS',
  HOME_PAGE = 'HOME_PAGE',
}